requirements.txt
pandas
requests
scikit-learn
numpy
PyYAML

